renamed files to match main component in file
