+++ dio/lib/adapter.dart
@@ -1,1 +1,2 @@
-export 'src/adapters/io_adapter.dart' show DefaultHttpClientAdapter;
+export 'src/adapters/default_http_client_adapter.dart'
+    show DefaultHttpClientAdapter;
